+++ relays/bin-substrate/src/chains/millau_messages_to_rialto.rs
@@ -16,7 +16,6 @@
 
 //! Millau-to-Rialto messages sync entrypoint.
 
-use messages_relay::relay_strategy::MixStrategy;
 use relay_millau_client::Millau;
 use relay_rialto_client::Rialto;
 use substrate_relay_helper::messages_lane::{
@@ -42,6 +41,4 @@ impl SubstrateMessageLane for MillauMessagesToRialto {
 		millau_runtime::Runtime,
 		millau_runtime::WithRialtoMessagesInstance,
 	>;
-
-	type RelayStrategy = MixStrategy;
 }
