+++ relays/bin-substrate/src/chains/rialto_messages_to_millau.rs
@@ -16,7 +16,6 @@
 
 //! Rialto-to-Millau messages sync entrypoint.
 
-use messages_relay::relay_strategy::MixStrategy;
 use relay_millau_client::Millau;
 use relay_rialto_client::Rialto;
 use substrate_relay_helper::messages_lane::{
@@ -42,6 +41,4 @@ impl SubstrateMessageLane for RialtoMessagesToMillau {
 		rialto_runtime::Runtime,
 		rialto_runtime::WithMillauMessagesInstance,
 	>;
-
-	type RelayStrategy = MixStrategy;
 }
