+++ relays/lib-substrate-relay/src/messages_source.rs
@@ -32,12 +32,10 @@ use async_trait::async_trait;
 use bp_messages::{
 	storage_keys::{operating_mode_key, outbound_lane_data_key},
 	InboundMessageDetails, LaneId, MessageNonce, MessagePayload, MessagesOperatingMode,
-	OutboundLaneData, OutboundMessageDetails, UnrewardedRelayersState,
+	OutboundLaneData, OutboundMessageDetails,
 };
 use bp_runtime::{messages::DispatchFeePayment, BasicOperatingMode, HeaderIdProvider};
-use bridge_runtime_common::messages::{
-	source::FromBridgedChainMessagesDeliveryProof, target::FromBridgedChainMessagesProof,
-};
+use bridge_runtime_common::messages::target::FromBridgedChainMessagesProof;
 use codec::{Decode, Encode};
 use frame_support::weights::Weight;
 use messages_relay::{
@@ -47,11 +45,11 @@ use messages_relay::{
 		SourceClientState,
 	},
 };
-use num_traits::{Bounded, Zero};
+use num_traits::Zero;
 use relay_substrate_client::{
-	AccountIdOf, AccountKeyPairOf, BalanceOf, BlockNumberOf, Chain, ChainWithMessages,
-	ChainWithTransactions, Client, Error as SubstrateError, HashOf, HeaderIdOf, IndexOf, SignParam,
-	TransactionEra, TransactionTracker, UnsignedTransaction,
+	AccountIdOf, AccountKeyPairOf, BalanceOf, BlockNumberOf, Chain, ChainWithMessages, Client,
+	Error as SubstrateError, HashOf, HeaderIdOf, IndexOf, SignParam, TransactionEra,
+	TransactionTracker, UnsignedTransaction,
 };
 use relay_utils::{relay_loop::Client as RelayClient, HeaderId};
 use sp_core::{Bytes, Pair};
@@ -368,39 +366,6 @@ where
 			target_to_source_headers_relay.require_more_headers(id.0).await;
 		}
 	}
-
-	async fn estimate_confirmation_transaction(
-		&self,
-	) -> <MessageLaneAdapter<P> as MessageLane>::SourceChainBalance {
-		let runtime_version = match self.source_client.runtime_version().await {
-			Ok(v) => v,
-			Err(_) => return BalanceOf::<P::SourceChain>::max_value(),
-		};
-		async {
-			let dummy_tx = P::SourceChain::sign_transaction(
-				SignParam::<P::SourceChain> {
-					spec_version: runtime_version.spec_version,
-					transaction_version: runtime_version.transaction_version,
-					genesis_hash: *self.source_client.genesis_hash(),
-					signer: self.transaction_params.signer.clone(),
-				},
-				make_messages_delivery_proof_transaction::<P>(
-					&self.transaction_params,
-					HeaderId(Default::default(), Default::default()),
-					Zero::zero(),
-					prepare_dummy_messages_delivery_proof::<P::SourceChain, P::TargetChain>(),
-					false,
-				)?,
-			)?
-			.encode();
-			self.source_client
-				.estimate_extrinsic_fee(Bytes(dummy_tx))
-				.await
-				.map(|fee| fee.inclusion_fee())
-		}
-		.await
-		.unwrap_or_else(|_| BalanceOf::<P::SourceChain>::max_value())
-	}
 }
 
 /// Ensure that the messages pallet at source chain is active.
@@ -437,30 +402,6 @@ fn make_messages_delivery_proof_transaction<P: SubstrateMessageLane>(
 		);
 	Ok(UnsignedTransaction::new(call.into(), transaction_nonce)
 		.era(TransactionEra::new(source_best_block_id, source_transaction_params.mortality)))
-}
-
-/// Prepare 'dummy' messages delivery proof that will compose the delivery confirmation transaction.
-///
-/// We don't care about proof actually being the valid proof, because its validity doesn't
-/// affect the call weight - we only care about its size.
-fn prepare_dummy_messages_delivery_proof<SC: Chain, TC: Chain>(
-) -> SubstrateMessagesDeliveryProof<TC> {
-	let single_message_confirmation_size =
-		bp_messages::InboundLaneData::<()>::encoded_size_hint_u32(1, 1);
-	let proof_size = TC::STORAGE_PROOF_OVERHEAD.saturating_add(single_message_confirmation_size);
-	(
-		UnrewardedRelayersState {
-			unrewarded_relayer_entries: 1,
-			messages_in_oldest_entry: 1,
-			total_messages: 1,
-			last_delivered_nonce: 1,
-		},
-		FromBridgedChainMessagesDeliveryProof {
-			bridged_header_hash: Default::default(),
-			storage_proof: vec![vec![0; proof_size as usize]],
-			lane: Default::default(),
-		},
-	)
 }
 
 /// Read best blocks from given client.
@@ -634,7 +575,6 @@ fn split_msgs_to_refine<Source: Chain + ChainWithMessages, Target: Chain>(
 mod tests {
 	use super::*;
 	use bp_runtime::Chain as ChainBase;
-	use codec::MaxEncodedLen;
 	use relay_rialto_client::Rialto;
 	use relay_rococo_client::Rococo;
 	use relay_wococo_client::Wococo;
@@ -698,19 +638,6 @@ mod tests {
 			validate_out_msgs_details::<Wococo>(&message_details_from_rpc(1..=5), 2..=5,),
 			Err(SubstrateError::Custom(_))
 		));
-	}
-
-	#[test]
-	fn prepare_dummy_messages_delivery_proof_works() {
-		let expected_minimal_size =
-			bp_wococo::AccountId::max_encoded_len() as u32 + Rococo::STORAGE_PROOF_OVERHEAD;
-		let dummy_proof = prepare_dummy_messages_delivery_proof::<Wococo, Rococo>();
-		assert!(
-			dummy_proof.1.encode().len() as u32 > expected_minimal_size,
-			"Expected proof size at least {}. Got: {}",
-			expected_minimal_size,
-			dummy_proof.1.encode().len(),
-		);
 	}
 
 	fn check_split_msgs_to_refine(
