+++ relays/messages/src/lib.rs
@@ -29,9 +29,9 @@ mod metrics;
 
 pub mod message_lane;
 pub mod message_lane_loop;
-pub mod relay_strategy;
 
 mod message_race_delivery;
+mod message_race_limits;
 mod message_race_loop;
 mod message_race_receiving;
 mod message_race_strategy;
